feat: Upgrade 5 — Create Session via panel + bug fixes
UPGRADE 5 — Interactive create-session flow
- New backend service `sessionCreationService` keeps an in-memory map of
  pending creations keyed by tempId with a 5-minute TTL reaper. Each entry
  owns a freshly-built GramJS TelegramClient against an empty StringSession.
- New REST surface mounted under /api/sessions:
    POST /create/start    — auth.SendCode, returns { tempId, codeType }
    POST /create/verify   — auth.SignIn; surfaces SESSION_PASSWORD_NEEDED
                            as { status: 'awaiting_password' }
    POST /create/password — client.checkPassword for 2FA accounts
    POST /create/resend   — auth.ResendCode, refreshes phoneCodeHash
    POST /create/cancel   — disconnects the temp client
    GET  /:id/download    — streams the encrypted session JSON file for
                            any session the user owns (works for sessions
                            uploaded via the panel and sessions created
                            via this new flow alike)
- On success the live, already-connected client is adopted into
  telegramService.clients so the heartbeat loop (Upgrade 1) takes over
  immediately and we do not reconnect twice in a row. The DB row is
  inserted with status=active, is_logged_in=true, keep_alive=true.
- Created sessions persist to <uploadDir>/<userId>/sessions/<uuid>.json
  in the same encrypted format the upload flow uses, so the downloaded
  file round-trips back through /upload.

Frontend
- New "Create Session" sidebar entry (UserPlus icon) at /create-session.
- 4-step stepper: Phone → OTP (with Resend / Cancel) → 2FA (only if
  Telegram demands it) → Done. The Done step shows the new session
  metadata and a "Download session file" button that calls the new
  GET /sessions/:id/download endpoint via a Blob → anchor click.
- Sessions list grows a per-row Download icon (blue-tinted) between View
  and Login/Logout, working for any session that has a file on disk.

Bug fixes uncovered while testing the v2 mega upgrade:

(1) Login deadlock during proxy assignment
    sessionService.loginSession holds SELECT … FOR UPDATE on the sessions
    row inside its own transaction; assignProxyForSession ran on a
    separate pool connection and (a) attempted to UPDATE that same row
    and (b) inserted into session_proxy_assignments which has an FK to
    sessions(id). Both required the row lock and waited forever on the
    outer transaction's release, which in turn waited for assignment to
    return → 30s Telegram-side timeout, login appeared to hang.
    Fix:
      • proxyService.assignProxyForSession no longer touches sessions
        directly — session_proxy_assignments is the canonical mapping.
      • loginSession now COMMITs the row lock, runs the assignment from
        outside the transaction, then BEGIN+re-locks for the rest of
        the login flow.

(2) OTP-job INSERT — "inconsistent types deduced for parameter $3"
    `($3 || ' seconds')::interval` with $3 used both as int and text
    confused Postgres' parameter inference. Fixed with explicit
    `$3::int` casts on both uses.

Smoke-tested validation paths:
  • Phone validation (missing / junk) returns the correct AppError codes.
  • Verify with bogus tempId returns FLOW_NOT_FOUND.
  • GET /sessions/4/download → 200 OK with attachment headers (existing
    +919914414217 session round-trips through download).

After (1) and (2) all four v2 tests pass:
  • Session restore: total=1 restored=1 failed=0 across backend restart.
  • Bulk Change-2FA with wrong password: success=0 failed=1, per-item
    drawer carries the underlying Telegram-side error.
  • Proxies: 19 free SOCKS5 proxies probe with real numeric latencies
    (133–488ms); a TEST-NET (198.51.100.42) manual proxy flips to
    is_working=false on first probe (8s timeout) and survives delete.

+++ backend/src/services/otpService.js
@@ -91,7 +91,8 @@ class OtpService {
       const insert = await client.query(
         `INSERT INTO otp_jobs (user_id, status, total_count, duration_seconds,
                                started_at, expires_at)
-         VALUES ($1, 'scanning', $2, $3, NOW(), NOW() + ($3 || ' seconds')::interval)
+         VALUES ($1, 'scanning', $2, $3::int, NOW(),
+                 NOW() + ($3::int || ' seconds')::interval)
          RETURNING id, created_at, started_at, expires_at`,
         [userId, ids.length, duration]
       );
